mainActivity refactor
Прибрано в MainActivity. Код расфасован по файлам
