+++ app/src/main/java/com/twendev/vulpes/lagopus/ui/NavigationManager.kt
@@ -1,0 +1,35 @@
+package com.twendev.vulpes.lagopus.ui
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.Stable
+import androidx.compose.runtime.State
+import androidx.compose.runtime.remember
+import androidx.navigation.NavBackStackEntry
+import androidx.navigation.NavHostController
+import androidx.navigation.NavOptionsBuilder
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+
+@Stable
+class NavigationManager(val navController : NavHostController) {
+    fun navTo(route: String, navOptionsBuilder: NavOptionsBuilder.() -> Unit) = navController.navigate(route, navOptionsBuilder)
+    fun navTo(route: String) {
+        navController.navigate(route)
+    }
+
+    fun navUp() {
+        navController.navigateUp()
+    }
+
+    @Composable
+    fun collectNavStackEntryAsState() : State<NavBackStackEntry?> {
+        return navController.currentBackStackEntryAsState()
+    }
+}
+
+@Composable
+fun rememberNavManager(navController: NavHostController = rememberNavController()) : NavigationManager {
+    return remember(Unit) {
+        NavigationManager(navController = navController)
+    }
+}
